Tratamento de mascara para o cnpj do Teatro

+++ DSW-T1/src/java/DAO/TeatroDAO.java
@@ -26,7 +26,7 @@ public class TeatroDAO extends GenericDAO{
             statement.setString(2, teatro.getSenha());
             statement.setString(3, teatro.getCidade());
             statement.setString(4, teatro.getNome());
-            statement.setInt(5, teatro.getCnpj());
+            statement.setLong(5, teatro.getCnpj());
             
             statement.executeUpdate();
             statement.close();
@@ -49,7 +49,7 @@ public class TeatroDAO extends GenericDAO{
                 String senha = resultSet.getString("senha");
                 String cidade = resultSet.getString("cidade");
                 String nome = resultSet.getString("nome");
-                Integer cnpj = resultSet.getInt("cnpj");
+                Long cnpj = resultSet.getLong("cnpj");
                 Integer id = resultSet.getInt("id");
                 
                 Teatro teatro = new Teatro(email, senha, cidade, nome, cnpj, id);
@@ -74,7 +74,7 @@ public class TeatroDAO extends GenericDAO{
             statement.setString(2, teatro.getSenha());
             statement.setString(3, teatro.getCidade());
             statement.setString(4, teatro.getNome());
-            statement.setInt(5, teatro.getCnpj());            
+            statement.setLong(5, teatro.getCnpj());            
             statement.setInt(6, teatro.getId());
 
             statement.executeUpdate();
@@ -113,7 +113,7 @@ public class TeatroDAO extends GenericDAO{
                 String senha = resultSet.getString("senha");
                 String cidade = resultSet.getString("cidade");
                 String nome = resultSet.getString("nome");
-                Integer cnpj = resultSet.getInt("cnpj");
+                Long cnpj = resultSet.getLong("cnpj");
                 Integer id = resultSet.getInt("id");
                 
                 Teatro teatro = new Teatro(email, senha, cidade, nome, cnpj, id);
@@ -142,7 +142,7 @@ public class TeatroDAO extends GenericDAO{
                 String senha = resultSet.getString("senha");
                 String cidade = resultSet.getString("cidade");
                 String nome = resultSet.getString("nome");
-                Integer cnpj = resultSet.getInt("cnpj");
+                Long cnpj = resultSet.getLong("cnpj");
                 
                 teatro = new Teatro(email, senha, cidade, nome, cnpj, id);
             }
